adds compilationMode to partial
