+++ packages/dashboards/src/lib/components/dashboard/dashboard.component.spec.ts
@@ -1,10 +1,11 @@
 import { SimpleChange, SimpleChanges } from "@angular/core";
 import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
 import { GridsterItem } from "angular-gridster2";
-import { IDashboard, IWidgets } from "src/lib/types";
+import { IDashboard } from "./types";
 
 import { NuiDashboardsModule } from "../../dashboards.module";
 import { WIDGET_POSITION_CHANGE } from "../../services/types";
+import { IWidgets } from "../widget/types";
 
 import { DashboardComponent } from "./dashboard.component";
 import { DEFAULT_GRIDSTER_CONFIG } from "./default-gridster-config";
