+++ packages/dashboards/src/lib/components/dashboard/public-api.ts
@@ -1,1 +1,2 @@
 export * from "./dashboard.component";
+export * from "./types";
