+++ packages/dashboards/src/lib/components/widget/public-api.ts
@@ -3,3 +3,4 @@ export * from "./widget-header/widget-header.component";
 export * from "./widget-body/widget-body.component";
 export * from "./widget-body-content/widget-body-content.component";
 export * from "./widget-header/types";
+export * from "./types";
