+++ packages/dashboards/src/lib/configurator/components/widget-editor/widget-editor.component.spec.ts
@@ -10,7 +10,7 @@ import { TemplateLoadErrorComponent } from "../../../components/template-load-er
 import { mockLoggerService } from "../../../mocks";
 import { NuiPizzagnaModule } from "../../../pizzagna/pizzagna.module";
 import { ProviderRegistryService } from "../../../services/provider-registry.service";
-import { IWidget, PizzagnaLayer } from "../../../types";
+import { PizzagnaLayer } from "../../../types";
 import { PreviewService } from "../../services/preview.service";
 import { ConfiguratorHeadingService } from "../../services/public-api";
 import { ConfiguratorComponent } from "../configurator/configurator.component";
@@ -18,6 +18,7 @@ import { ConfiguratorHeadingComponent } from "../heading/configurator-heading.co
 import { DashwizStepComponent } from "../wizard/dashwiz-step/dashwiz-step.component";
 import { DashwizButtonsComponent } from "../wizard/dashwiz/dashwiz-buttons.component";
 import { DashwizComponent } from "../wizard/dashwiz/dashwiz.component";
+import { IWidget } from "../../../components/widget/types";
 
 import { WidgetEditorComponent } from "./widget-editor.component";
