+++ packages/dashboards/src/lib/configurator/components/wizard/dashwiz/public-api.ts
@@ -1,2 +1,3 @@
 export * from "./dashwiz.component";
 export * from "./dashwiz-buttons.component";
+export * from "./model";
