+++ packages/dashboards/src/lib/widget-types/common/widget/components.ts
@@ -4,7 +4,8 @@ import { WidgetBodyContentComponent } from "../../../components/widget/widget-bo
 import { WidgetBodyComponent } from "../../../components/widget/widget-body/widget-body.component";
 import { WidgetHeaderComponent } from "../../../components/widget/widget-header/widget-header.component";
 import { NOVA_LOADING_ADAPTER, NOVA_STATUS_CONTENT_FALLBACK_ADAPTER } from "../../../services/types";
-import { HttpStatusCode, IComponentConfiguration, IProviderConfiguration, IWidgetErrorDisplayProperties, WellKnownProviders } from "../../../types";
+import { HttpStatusCode, IComponentConfiguration, IProviderConfiguration, WellKnownProviders } from "../../../types";
+import { IWidgetErrorDisplayProperties } from "../../../components/widget/types";
 
 import { ErrorNodeKey } from "./types";
