+++ packages/dashboards/src/lib/widget-types/common/widget/legend.ts
@@ -1,0 +1,10 @@
+export enum LegendPlacement {
+  None = "None",
+  Right = "Right",
+  Bottom = "Bottom",
+}
+
+export interface ILegendPlacementOption {
+  id: LegendPlacement;
+  label: string;
+}
