+++ packages/dashboards/src/lib/widget-types/common/widget/public-api.ts
@@ -1,3 +1,4 @@
 export * from "./components";
 export * from "./providers";
 export * from "./types";
+export * from "./legend";
