+++ packages/dashboards/src/lib/widget-types/embedded-content/embedded-content.ts
@@ -1,5 +1,6 @@
 import { DEFAULT_PIZZAGNA_ROOT } from "../../services/types";
-import { IWidgetTypeDefinition, WellKnownPathKey } from "../../types";
+import { WellKnownPathKey } from "../../types";
+import { IWidgetTypeDefinition } from "../../components/widget/types";
 
 import { embeddedContentConfigurator } from "./embedded-content-configurator";
 import { embeddedContentWidget } from "./embedded-content-widget";
