+++ packages/dashboards/src/lib/widget-types/table/table.ts
@@ -1,5 +1,6 @@
 import { DEFAULT_PIZZAGNA_ROOT } from "../../services/types";
-import { IWidgetTypeDefinition, WellKnownPathKey } from "../../types";
+import { WellKnownPathKey } from "../../types";
+import { IWidgetTypeDefinition } from "../../components/widget/types";
 
 import { tableConfigurator } from "./table-configurator";
 import { tableWidget } from "./table-widget";
