+++ packages/dashboards/src/lib/widget-types/timeseries/timeseries.ts
@@ -1,5 +1,6 @@
 import { DEFAULT_PIZZAGNA_ROOT } from "../../services/types";
-import { IWidgetTypeDefinition, WellKnownPathKey } from "../../types";
+import { WellKnownPathKey } from "../../types";
+import { IWidgetTypeDefinition } from "../../components/widget/types";
 
 import { timeseriesConfigurator } from "./timeseries-configurator";
 import { timeseriesWidget } from "./timeseries-widget";
